Add concrete type info to generic Flink interfaces
Java to the rescue!

Add a Java class to be a simple data-structure to hold our word-count
data.

Define an interface in Java which applies concrete types to generic
Flink interfaces.

Reify'ing this *now* has all the type information needed by Flink!
Clojure code for the win, with a little help from Java.