Convert the WriteIntoKafka example into Clojure
This one has an important and slightly discouraging discovery. We
implement our own source here which keeps returning new strings every
second. We need to implement the `SourceFunction` interface for this.

Initially, I thought that the previous approach of implementing a
concrete Interface which extends `SourceFunction<String>` would be
enough, but this is not true. This is because of the way that the two
methods of the `SourceFunction` Interface interact with each other. The
`run` method starts a possible infinite loop, and the `cancel` method
stops the loop started by the `run` method. This requires us to mutate
state held within the Class, but without using any of the Clojure state
management methods like atom, ref etc. This is because these are not
`Serializable`. I don't know how to do this from Clojure, and I'm 99%
sure that it is not possible.

Due to this, I had to write a class in Java which implemented the
`SourceFunction` interface.